requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
Flask-WTF==1.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
